move to metrics.go
