+++ main.go
@@ -1,26 +1,18 @@
 package main
 
 import (
-	"encoding/json"
-	"github.com/hashicorp/vault/api"
 	"github.com/pkg/errors"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
 	"log"
 	"net/http"
 	"os"
-	"time"
 )
 
 var Logger *log.Logger
-var tokenTTL = prometheus.NewGauge(prometheus.GaugeOpts{
-	Namespace: "vault_token_renewer",
-	Subsystem: "token",
-	Name:      "ttl",
-})
 
 func init() {
-	prometheus.MustRegister(tokenTTL)
+	prometheus.MustRegister(tokenTTLCollector)
 }
 
 func main() {
@@ -52,20 +44,4 @@ func run() error {
 	go http.ListenAndServe(":8080", nil)
 
 	return tokenRenewer.Run()
-}
-
-func getTokenTTL(vault *api.Client) error {
-	for {
-		secret, err := vault.Auth().Token().Lookup(os.Getenv("VAULT_TOKEN"))
-		if err != nil {
-			return err
-		}
-		ttl, err := secret.Data["ttl"].(json.Number).Float64()
-		if err != nil {
-			return err
-		}
-		tokenTTL.Set(ttl)
-		time.Sleep(10 * time.Second)
-	}
-	return nil
 }
